move components into modular functions
